# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: "The Monika Cloud Cloud Platform API provides web services for the Monika suite of apps and systems."
  version: "0.1.0"
  title: "Monika Cloud Patform API"
  termsOfService: "https://cloud.monika.com/terms/"
  contact:
    email: "swood@monika.com"
  license:
    name: "Commercial"
    url: "http://cloud.monika.com/licenses/license-1.0.html"
host: "127.0.0.1:9000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /user:
    x-swagger-router-controller: user.controller
    post:
      tags:
      - "user"
      summary: "Add a new user to the system"
      description: "Add a new user to the system"
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "The user object to be added"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "User successfuly added"
          schema:
            type: object
            properties:
              userId:
                type: string
                format: uuid
        400:
          description: "The user data posted was not valid"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "There is a user already matching that data"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
...
  /user/{userId}/permission:
    x-swagger-router-controller: userPermission.controller
    post:
      tags:
      - "user"
      summary: "Add a new permission for a user"
      description: "Add a new permission for a user"
      operationId: "addPermissionForUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who should have this permission added to their list of permissions"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "The permission object to be added"
        required: true
        schema:
          type: array
          items:
            $ref: "#/definitions/Permission"
      responses:
        200:
          description: "Permission successfuly added"
          schema:
            type: string
        404:
          description: "User not found"
        405:
          description: "Invalid input"
    put:
      tags:
      - "user"
      summary: "Update an existing permission for a user"
      description: "Updates the permission details"
      operationId: "updatePermissionForUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's permission is being updated"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "The permission object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Permission"
      responses:
        200:
          description: "Permission successfuly updated"
        405:
          description: "Invalid input"
        404:
          description: "User or Permission not found"
  /user/{userId}/permission/list:
    x-swagger-router-controller: userPermission.controller
    post:
      tags:
      - "user"
      summary: "Get a list of permissions for a user, filtered and sorted"
      description: "Get a list of permissions for a user, filtered and sorted."
      operationId: "listPermissionsForUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's permissions you want to list"
        required: true
        type: string
        format: uuid
      - in: body
        name: body
        description: "Parameters that filter and sort the list of permissions"
        required: false
        schema:
          type: object
          properties:
            permissionListFilterQuery:
              $ref: "#/definitions/PermissionListFilterQuery"
            permissionListSortQuery:
              $ref: "#/definitions/PermissionListSortQuery"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Permission"
        400:
          description: "Invalid query data"
  /user/{userId}/permission/{permissionId}:
    x-swagger-router-controller: userPermission.controller
    get:
      tags:
      - "user"
      summary: "Retreive the details of a users permission by permission id"
      description: "Retreive the details of a users permission by permission id"
      operationId: "getPermissionForUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user account you would like to activate"
        required: true
        type: string
        format: uuid
      - name: "permissionId"
        in: path
        description: "The id of the permission you would like to get the details of"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Permission"
        404:
          description: "Permission not found"
    delete:
      tags:
      - "permission"
      summary: "Deletes a user's permission"
      description: "Deletes or removes a user's permission"
      operationId: "deletePermissionForUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: path
        description: "The id of the user who's permission you want to delete or remove"
        required: true
        type: string
        format: uuid
      - name: "permissionId"
        in: path
        description: "The id of the permission you would like to delete or remove"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "User's permission was deleted."
        404:
          description: "Permission not found"
  /user/permissionType/list:
    x-swagger-router-controller: userPermission.controller
    get:
      tags:
      - "permission"
      summary: "Get a list of permission types"
      description: "Get a list of permissions types"
      operationId: "listUserPermissionTypes"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/PermissionType"
        400:
          description: "Invalid query data"
  /user/permissionType/{permissionTypeId}:
    x-swagger-router-controller: userPermission.controller
    get:
      tags:
      - "permission"
      summary: "Retreive the details of the permission type by id"
      description: "Retreive the details of the permission type by id"
      operationId: "getUserPermissionType"
      produces:
      - "application/json"
      parameters:
      - name: "permissionTypeId"
        in: path
        description: "The id of the permission type you would like to retrieve"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/PermissionType"
        404:
          description: "PermissionType not found"
...
  /tenant/{tenantId}:
    x-swagger-router-controller: tenant.controller
    get:
      tags:
      - "tenant"
      summary: "Retreive the details of the tenant by id"
      description: "Retreive the details of the tenant by id"
      operationId: "getTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to activate"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Tenant"
        400:
          description: "Tenant already active"
        404:
          description: "Tenant not found"
    patch:
      tags:
      - "tenant"
      summary: "Update an existing tenant's details in the system"
      description: "Updates the tenants details"
      operationId: "updateTenant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to update"
        required: true
        type: string
        format: uuid
      - name: JsonPatch
        in: body
        required: true
        schema:
          $ref: "#/definitions/PatchRequest"
      responses:
        200:
          description: "Tenant successfuly updated"
          schema:
            type: string
        404:
          description: "Tenant not found"
        405:
          description: "Invalid input"      
    delete:
      x-swagger-router-controller: tenant.controller
      tags:
      - "tenant"
      summary: "Delete the tenant by id"
      description: "Delete the tenant by id. Note, this is a permanent operation which cannot be undone. Make a backup first."
      operationId: "deleteTenant"
      produces:
      - "application/json"
      parameters:
      - name: "tenantId"
        in: path
        description: "The id of the tenant you would like to delete"
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: "successful operation, tenant deleted"
          schema:
            type: string
        404:
          description: "Tenant not found"
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
  Permission:
    type: object
    properties:
      id:
        type: string
        format: uuid
      permissionTypeId:
        type: string
        format: uuid
      tenantId:
        type: string
        format: uuid
      regionId:
        type: string
        format: uuid
      siteId:
        type: string
        format: uuid
  PermissionType  :
    type: object
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
      description:
        type: string
      tenantScope:
        type: boolean
        description: "Whether this permission can be granted with tenant scope"
      regionScope:
        type: boolean
        description: "Whether this permission can be granted with region scope"
      siteScope:
        type: boolean
        description: "Whether this permission can be granted with site scope"
  PermissionListSortQuery:
    type: object
    properties:
      sortBy:
        type: array
        items:
          type: string
          description: "Sort By"
          enum:
          - "name"
          - "date created"
      order:
        type: string
        description: "Sort Order"
        enum:
        - "ascending"
        - "descending"
  PermissionListFilterQuery:
    type: object
    properties:
      userId:
        type: string
        format: uuid
      keyword:
        type: string
        description: "perform a keyword search and only show those results with matches"
  Tenant:
    type: object
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
        description: "Typically the trading name of the business, e.g. IKEA"
      legalName:
        type: string
        description: "The legal name of the company, e.g. IKEA Pty Ltd"
      email:
        type: string
        description: "Company Email e.g admin@ikea.com"
      phone:
        type: string
        description: "Company phone number e.g. 0909090909"  
      address:
        type: string
        description: "Company address e.g. Unit 149 block 45 " 
      city:
        type: string
        description: "Company City e.g. Melbourne City" 
      country:
        type: string
        description: "Company country e.g. Australia"    
      databaseVersion:
        type: string
      databaseSizeBytes:
        type: integer
        format: int64
      subdomain:
        type: string
      created:
        type: string
        format: date-time
      lastModified:
        type: string
        format: date-time
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      profilePic:
        type: string
        format : file
...
  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"
  PatchDocument: 
    description: A JSONPatch document as defined by RFC 6902 
    required:
     - "op"
     - "path"
    properties: 
     op: 
      type: string 
      description: The operation to be performed 
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path: 
      type: string 
      description: A JSON-Pointer 
     value: 
      type: string 
      description: The value to be used within the operations.
     from: 
      type: string 
      description: A string containing a JSON Pointer value.
  User:
    type: object
    required:
    - "email"
    - "ownership"
    properties:
      id:
        type: string
        format: uuid
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
        format: "email"
      password:
        type: string
      phone:
        type: string
      status:
        type: string
        description: "Status"
        enum:
        - "active"
        - "deactivated"
      companyName:
        type: string
      profilePic:
        type: string
        format : file
      roleId:
        type: string
        format: uuid
      roleName:
        type: string
      ownership:
        type: object
        properties:
          ownershipType:
            type: string
            description: "What type of object the user account is owned by"
            enum:
            - "monika"
            - "none"
            - "tenant"
            - "region"
            - "site"
          tenantId:
            type: string
            format: uuid
          tenantName:
            type: string
          regionId:
            type: string
            format: uuid
          regionName:
            type: string
          siteId:
            type: string
            format: uuid
          siteName:
            type: string
      created:
        type: string
        format: date-time
      lastModified:
        type: string
        format: date-time
...